ci: fix dev-ci lint and TestNotify failures
- gofmt all files to satisfy goimports linter
- Remove redundant uint64/StorageSize conversions (unconvert)
- Trim unnecessary leading/trailing newlines (whitespace)
- .golangci.yml: skip generated ABI files and exclude unused/staticcheck
  for Metadium-specific paths (intentionally retained code)
- rpc/subscription_test.go: update TestNotify fixture for Camellia
  header fields (withdrawalsRoot, parentBeaconBlockRoot)

+++ core/blockchain.go
@@ -2055,7 +2055,7 @@ func (bc *BlockChain) insertSideChain(block *types.Block, it *insertIterator) (i
 		block := bc.GetBlock(hashes[i], numbers[i])
 
 		blocks = append(blocks, block)
-		memory += uint64(block.Size())
+		memory += block.Size()
 
 		// If memory use grew too large, import and continue. Sadly we need to discard
 		// all raised events and logs from notifications since we're too heavy on the
